requirements.txt
python = "^3.10"
pytsk3 = "^20230125"
libewf-python = "^20230212"
pyyaml = "^6.0"
argparse = "^1.4.0"